requirements.txt
# Core ML/Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Technical Analysis
talib>=0.4.25

# MT5 Trading
MetaTrader5>=5.0.37

# Dashboard
dash>=2.14.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0

# Optional: Advanced ML (uncomment if needed)
# tensorflow>=2.13.0
# torch>=2.0.0

# Optional: Advanced TA (uncomment if needed)
# ta>=0.10.2